+++ src/__tests__/createLoggerMiddleware.test.js
@@ -5,10 +5,10 @@ import {it, vi, expect, describe, beforeEach, afterEach} from 'vitest';
 import Logger from '../Logger';
 import createLoggerMiddleware from '../createLoggerMiddleware';
 
-function createServer(logger) {
+function createServer(logger, {maxGraphQLVariablesLength} = {}) {
   const server = express();
 
-  server.use(createLoggerMiddleware({logger}));
+  server.use(createLoggerMiddleware({logger, maxGraphQLVariablesLength}));
   server.get('/', (req, res) => res.json({success: true}));
   server.get('/500', () => {
     throw new Error('500');
@@ -166,6 +166,65 @@ describe('createLoggerMiddleware', () => {
       data: null
     });
     expect(stderrCalls[1].level).toBe('ERROR');
+  });
+
+  it('reduces the log payload of `variables` to the desired maxGraphQLVariablesLength', async () => {
+    const logger = new Logger({service: 'pizza-shop', isProduction: true});
+    const server = createServer(logger, {maxGraphQLVariablesLength: 12});
+
+    await supertest(server).get('/');
+    await supertest(server).get('/404');
+
+    await supertest(server).post('/graphql').send(mockGraphQLPayload);
+
+    const stdoutCalls = console._stdout.write.mock.calls.map((call) =>
+      JSON.parse(call[0])
+    );
+
+    expect(stdoutCalls[2].level).toBe('DEBUG');
+    expect(stdoutCalls[2].meta.graphql).toEqual({
+      operationName: 'createPizza',
+      variables: '{"pizza":{"t […] max payload length reached (12 chars)'
+    });
+  });
+
+  it('does not log `variables` when maxGraphQLVariablesLength = 0', async () => {
+    const logger = new Logger({service: 'pizza-shop', isProduction: true});
+    const server = createServer(logger, {maxGraphQLVariablesLength: 0});
+
+    await supertest(server).get('/');
+    await supertest(server).get('/404');
+
+    await supertest(server).post('/graphql').send(mockGraphQLPayload);
+
+    const stdoutCalls = console._stdout.write.mock.calls.map((call) =>
+      JSON.parse(call[0])
+    );
+
+    expect(stdoutCalls[2].level).toBe('DEBUG');
+    expect(stdoutCalls[2].meta.graphql).toEqual({
+      operationName: 'createPizza'
+    });
+  });
+
+  it('logs all `variables` when maxGraphQLVariablesLength = -1', async () => {
+    const logger = new Logger({service: 'pizza-shop', isProduction: true});
+    const server = createServer(logger, {maxGraphQLVariablesLength: -1});
+
+    await supertest(server).get('/');
+    await supertest(server).get('/404');
+
+    await supertest(server).post('/graphql').send(mockGraphQLPayload);
+
+    const stdoutCalls = console._stdout.write.mock.calls.map((call) =>
+      JSON.parse(call[0])
+    );
+
+    expect(stdoutCalls[2].level).toBe('DEBUG');
+    expect(stdoutCalls[2].meta.graphql).toEqual({
+      operationName: 'createPizza',
+      variables: {pizza: {toppings: ['salami']}}
+    });
   });
 
   it('masks confidential headers', async () => {